+++ packages/optimizely-sdk/lib/utils/enums/index.ts
@@ -304,12 +304,3 @@ export enum ODP_USER_KEY {
   VUID = 'vuid',
   FS_USER_ID = 'fs_user_id',
 }
-
-/**
- * Possible states of ODP integration
- */
-export enum ODP_CONFIG_STATE {
-  UNDETERMINED = 0,
-  INTEGRATED,
-  NOT_INTEGRATED = 2,
-}
